added vlc-jjk: ugly hack so DPMS (screensaver) gets enabled/disabled on pause/resume
Note: this is not a good way to do this. the proper way would have been to trigger the Inhibit call for modules with capability "inhibit" on pause/resume, but I'm not sure how to do that from window.c or how to trigger the calls from elsewhere, so I went this ugly way...
Also, it would have been better to remove inhibit/xdg.c and replace it with this, but then it messes modules and stuff and it gets a whole lot more complicated, and I didn't have time to do things nicer. Plus, remember this is not the right way to do it anyways, so I just wanted to fix this for me, so my screen don't turn off in the middle of a video anymore (which I can now ensure won't happen by just pausing/resuming to be sure)