tt-rss-jjk: Update to work with open_basedir on
Now that /usr/share/webapps is in open_basedir, make tt-rss work with it. The
patch just removed the check (refusing to work w/ open_basedir), and what it
means is that instead of using cURL simple PHP's file_get_contents() will be
used. It shouldn't really be a problem though.