cryptsetup-jjk: Workaround slowness due to libgcrypt 1.6.0
Not sure why but using libgcrypt's pbkdf2 is much slower than cryptsetup's
internal one (which was used until then due to a (now fixed) bug in libgcrypt).

This is 2s vs 2m kinda slow when luksOpenning a device with a keyfile, so it is
quite noticable.

cryptsetup allows to disable using libgcrypt's implementation, but not force it
off - so w/ 1.6.0 it will always be on. Small patch to allow disabling it, as a
workaround.