firefox-jjk: Check sha512 from mozilla's signed SHA512SUMS
Since they don't sign the source code anymore, let's still find a way to verify
the signature. They sign the list of hashes, e.g. SHA512SUMS, so we include it
and its signature as source (therefore makepkg will verify said signature) and
then in prepare we just have an ugly hack to get & check the sha512 for the
source code.